Portfolio/requirements.txt
numpy==1.23.5
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
transformers==4.44.2
torch==2.4.0
nltk==3.9.1
tqdm==4.66.5
argparse==1.4.0
datasets==2.21.0
spacy==3.7.6
accelerate==0.34.2
peft==0.12.0
python-dotenv==1.0.1